requirements-poc.txt
# PoC UI + PDF + Gemini orchestration (use alongside Chroma / sentence-transformers stack)
streamlit>=1.28.0
fpdf2>=2.7.0
google-genai>=1.0.0
